requirements/dev.txt
-r base.txt

django-debug-toolbar==0.11.0
mock==1.0.1
factory_boy==2.3.1
